Refactor extension method.
